Use Moshi Codegen correctly
