+++ app/src/main/java/com/jshvarts/healthreads/domain/Book.kt
@@ -1,0 +1,16 @@
+package com.jshvarts.healthreads.domain
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+@Entity(tableName = "book")
+data class Book(
+  @PrimaryKey
+  @Json(name = "primary_isbn10") val isbn: String,
+  val title: String,
+  @Json(name = "book_image") val imageUrl: String,
+  val contributor: String
+)
