+++ app/src/main/java/com/jshvarts/healthreads/domain/BooksJsonAdapter.kt
@@ -1,37 +1,28 @@
 package com.jshvarts.healthreads.domain
 
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
 import com.squareup.moshi.FromJson
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import com.squareup.moshi.JsonQualifier
 import com.squareup.moshi.ToJson
 
 /**
+ * We are using Moshi Codegen and rely on custom adapters to parse json.
+ *
  * Json payload is as follows:
  *
- * results {
- * --> books [
- * ----> { book1 }
- * ----> { book2 }
- * ----> { bookN }
+ * results { -> BookResults
+ * --> books [ -> BookList
+ * ----> { book1 } -> Book
+ * ----> { book2 } -> Book
+ * ----> { bookN } -> Book
  */
-@JsonClass(generateAdapter = true)
-@Entity(tableName = "book")
-data class Book(
-  @PrimaryKey
-  @Json(name = "primary_isbn10") val isbn: String,
-  val title: String,
-  @Json(name = "book_image") val imageUrl: String,
-  val contributor: String
-)
 
+@JsonClass(generateAdapter = true)
 data class BookResults(
   val results: BookList
 )
 
+@JsonClass(generateAdapter = true)
 data class BookList(
   val books: List<Book>
 )
